moved deletion of blob code to teardown

+++ src/functionalTest/java/uk/gov/hmcts/reform/locationrefdata/cameltest/LrdApplicationTest.java
@@ -121,7 +121,6 @@ class LrdApplicationTest extends LrdIntegrationBaseTest {
                 .ccdServiceName("ccd-service2").serviceCode("AAA2").build()
         ), 4);
         validateLrdServiceFileAudit(jdbcTemplate, auditSchedulerQuery, "Success", UPLOAD_ORG_SERVICE_FILE_NAME);
-        lrdBlobSupport.deleteBlob(UPLOAD_ORG_SERVICE_FILE_NAME);
     }
 
     private void testInsertion() throws Exception {
@@ -145,8 +144,6 @@ class LrdApplicationTest extends LrdIntegrationBaseTest {
         ), 4);
         //Validates Success Audit
         validateLrdServiceFileAudit(jdbcTemplate, auditSchedulerQuery, "Success", UPLOAD_ORG_SERVICE_FILE_NAME);
-        //Delete Uploaded test file with Snapshot delete
-        lrdBlobSupport.deleteBlob(UPLOAD_ORG_SERVICE_FILE_NAME);
     }
 
     @Test
@@ -196,7 +193,6 @@ class LrdApplicationTest extends LrdIntegrationBaseTest {
         List<Map<String, Object>> auditDetailsNextRun = jdbcTemplate.queryForList(auditSchedulerQuery);
         final Timestamp timestampNextRun = (Timestamp) auditDetailsNextRun.get(0).get("scheduler_end_time");
         assertEquals(timestamp, timestampNextRun);
-        lrdBlobSupport.deleteBlob(UPLOAD_ORG_SERVICE_FILE_NAME);
     }
 
     @Test
@@ -228,13 +224,13 @@ class LrdApplicationTest extends LrdIntegrationBaseTest {
         ), 6);
         //Validates Success Audit
         validateLrdServiceFileAudit(jdbcTemplate, auditSchedulerQuery, "Success", UPLOAD_ORG_SERVICE_FILE_NAME);
-        //Delete Uploaded test file with Snapshot delete
-        lrdBlobSupport.deleteBlob(UPLOAD_ORG_SERVICE_FILE_NAME);
     }
 
     @AfterEach
     void tearDown() throws Exception {
         //Delete Uploaded test file with Snapshot delete
-        lrdBlobSupport.deleteBlob(UPLOAD_ORG_SERVICE_FILE_NAME);
+        if (lrdBlobSupport.isBlobPresent(UPLOAD_ORG_SERVICE_FILE_NAME)) {
+            lrdBlobSupport.deleteBlob(UPLOAD_ORG_SERVICE_FILE_NAME);
+        }
     }
 }
