+++ src/functionalTest/java/uk/gov/hmcts/reform/locationrefdata/cameltest/LrdBuildingLocationFileStatusCheck.java
@@ -112,7 +112,6 @@ public class LrdBuildingLocationFileStatusCheck extends LrdIntegrationBaseTest {
         Assertions.assertEquals(1, jdbcTemplate.queryForList(lrdAuditSqlFailure).size());
         List<Map<String, Object>> buildingLocations = jdbcTemplate.queryForList(lrdBuildingLocationSelectQuery);
         assertFalse(buildingLocations.isEmpty());
-        deleteFile();
     }
 
     private void deleteFile() throws Exception {
@@ -174,8 +173,9 @@ public class LrdBuildingLocationFileStatusCheck extends LrdIntegrationBaseTest {
     @AfterEach
     void tearDown() throws Exception {
         //Delete Uploaded test file with Snapshot delete
-        lrdBlobSupport.deleteBlob(UPLOAD_FILE_NAME);
+        if (lrdBlobSupport.isBlobPresent(UPLOAD_ORG_SERVICE_FILE_NAME)) {
+            lrdBlobSupport.deleteBlob(UPLOAD_FILE_NAME);
+        }
     }
-
 
 }
